But fixes, adjustments for 64-bits and VS 2017

+++ include/id3/id3lib_frame.h
@@ -78,7 +78,7 @@ public:
 
   ID3_Field*  GetField(ID3_FieldID name) const;
 
-  size_t      NumFields() const;
+  uint32      NumFields() const;
 
   const char* GetDescription() const;
   static const char* GetDescription(ID3_FrameID);
